backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Martin Sit <46298951+martin226@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
requests = "^2.27.1"
PyYAML = "^6.0"
ua-parser = "^0.10.0"
user-agents = "^2.2.0"
django-user-agents = "^0.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
